+++ sdk/keyvault/azure-security-keyvault-keys/src/key_client.cpp
@@ -344,7 +344,7 @@ Azure::Response<KeyVaultKey> KeyClient::UpdateKeyProperties(
       {_detail::KeysPath, properties.Name, properties.Version});
 }
 
-Azure::Response<std::vector<uint8_t>> KeyClient::BackupKey(
+Azure::Response<Azure::Security::KeyVault::Keys::BackupKeyResult> KeyClient::BackupKey(
     std::string const& name,
     Azure::Core::Context const& context) const
 {
@@ -358,8 +358,9 @@ Azure::Response<std::vector<uint8_t>> KeyClient::BackupKey(
       {_detail::KeysPath, name, "backup"});
 
   // Convert the internal KeyBackup model to a raw vector<uint8_t>.
-  return Azure::Response<std::vector<uint8_t>>(
-      response.Value.Value, std::move(response.RawResponse));
+  return Azure::Response<Azure::Security::KeyVault::Keys::BackupKeyResult>(
+      Azure::Security::KeyVault::Keys::BackupKeyResult{response.Value.Value},
+      std::move(response.RawResponse));
 }
 
 Azure::Response<KeyVaultKey> KeyClient::RestoreKeyBackup(